Call testing jobs in checkFineTuneStatus task
