+++ app/src/server/tasks/fineTuning/checkFineTuneStatus.task.ts
@@ -1,6 +1,11 @@
+import { type ChatCompletionMessage } from "openai/resources/chat";
+
 import { prisma } from "~/server/db";
 import defineTask from "../defineTask";
 import { trainingStatus } from "~/utils/modal";
+import { pruneInputMessages } from "~/modelProviders/fine-tuned/getCompletion";
+import { countLlamaChatTokens } from "~/utils/countTokens";
+import { queueGetTestResult } from "../getTestResult.task";
 
 const runOnce = async () => {
   const trainingJobs = await prisma.fineTune.findMany({
@@ -22,7 +27,40 @@ const runOnce = async () => {
       try {
         const resp = await trainingStatus({ callId: job.modalTrainingJobId });
         if (resp.status === "done") {
-          // TODO: kick off job to run test set evals here
+          const fineTune = await prisma.fineTune.findUnique({
+            where: { id: job.id },
+          });
+          if (!fineTune) return;
+          const pruningRules = await prisma.pruningRule.findMany({
+            where: { fineTuneId: fineTune.id },
+            select: { textToMatch: true },
+          });
+          const stringsToPrune = pruningRules.map((rule) => rule.textToMatch);
+          const datasetEntries = await prisma.datasetEntry.findMany({
+            where: { datasetId: fineTune.datasetId, outdated: false, type: "TEST" },
+            select: { id: true, input: true },
+            orderBy: { sortKey: "desc" },
+          });
+          // create fineTuneTestEntry for each dataset entry
+          await prisma.fineTuneTestingEntry.createMany({
+            data: datasetEntries.map((entry) => {
+              const prunedInput = pruneInputMessages(
+                entry.input as unknown as ChatCompletionMessage[],
+                stringsToPrune,
+              );
+              const prunedInputTokens = countLlamaChatTokens(prunedInput);
+              return {
+                fineTuneId: fineTune.id,
+                datasetEntryId: entry.id,
+                prunedInputTokens,
+                prunedInput,
+              };
+            }),
+            skipDuplicates: true,
+          });
+          for (const entry of datasetEntries) {
+            await queueGetTestResult(fineTune.id, entry.id);
+          }
 
           await prisma.fineTune.update({
             where: { id: job.id },
