+++ trainer/openapi.json
@@ -38,7 +38,10 @@
                 "schema": {
                   "type": "object",
                   "properties": {
-                    "trainingBlobDownloadUrl": {
+                    "trainingDataUrl": {
+                      "type": "string"
+                    },
+                    "huggingFaceModelId": {
                       "type": "string"
                     },
                     "baseModel": {
@@ -52,89 +55,9 @@
                     }
                   },
                   "required": [
-                    "trainingBlobDownloadUrl",
+                    "trainingDataUrl",
+                    "huggingFaceModelId",
                     "baseModel"
-                  ],
-                  "additionalProperties": false
-                }
-              }
-            }
-          },
-          "default": {
-            "$ref": "#/components/responses/error"
-          }
-        }
-      }
-    },
-    "/status-update": {
-      "post": {
-        "operationId": "recordStatusUpdate",
-        "description": "Record a status update for a model",
-        "security": [
-          {
-            "Authorization": []
-          }
-        ],
-        "requestBody": {
-          "required": true,
-          "content": {
-            "application/json": {
-              "schema": {
-                "type": "object",
-                "properties": {
-                  "fineTuneId": {
-                    "type": "string"
-                  },
-                  "status": {
-                    "type": "string",
-                    "enum": [
-                      "UPLOADING_DATASET",
-                      "PENDING",
-                      "TRAINING",
-                      "AWAITING_DEPLOYMENT",
-                      "DEPLOYING",
-                      "DEPLOYED",
-                      "ERROR"
-                    ]
-                  }
-                },
-                "required": [
-                  "fineTuneId",
-                  "status"
-                ],
-                "additionalProperties": false
-              }
-            }
-          }
-        },
-        "parameters": [],
-        "responses": {
-          "200": {
-            "description": "Successful response",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "type": "object",
-                  "properties": {
-                    "status": {
-                      "anyOf": [
-                        {
-                          "type": "string",
-                          "enum": [
-                            "ok"
-                          ]
-                        },
-                        {
-                          "type": "string",
-                          "enum": [
-                            "error"
-                          ]
-                        }
-                      ]
-                    }
-                  },
-                  "required": [
-                    "status"
                   ],
                   "additionalProperties": false
                 }
