+++ trainer/trainer/api_client/models/__init__.py
@@ -2,18 +2,8 @@
 
 from .get_training_info_response_200 import GetTrainingInfoResponse200
 from .get_training_info_response_200_base_model import GetTrainingInfoResponse200BaseModel
-from .record_status_update_json_body import RecordStatusUpdateJsonBody
-from .record_status_update_json_body_status import RecordStatusUpdateJsonBodyStatus
-from .record_status_update_response_200 import RecordStatusUpdateResponse200
-from .record_status_update_response_200_status_type_0 import RecordStatusUpdateResponse200StatusType0
-from .record_status_update_response_200_status_type_1 import RecordStatusUpdateResponse200StatusType1
 
 __all__ = (
     "GetTrainingInfoResponse200",
     "GetTrainingInfoResponse200BaseModel",
-    "RecordStatusUpdateJsonBody",
-    "RecordStatusUpdateJsonBodyStatus",
-    "RecordStatusUpdateResponse200",
-    "RecordStatusUpdateResponse200StatusType0",
-    "RecordStatusUpdateResponse200StatusType1",
 )
