--- trainer/trainer/api_client/models/record_status_update_json_body_status.py
@@ -1,14 +1,0 @@
-from enum import Enum
-
-
-class RecordStatusUpdateJsonBodyStatus(str, Enum):
-    AWAITING_DEPLOYMENT = "AWAITING_DEPLOYMENT"
-    DEPLOYED = "DEPLOYED"
-    DEPLOYING = "DEPLOYING"
-    ERROR = "ERROR"
-    PENDING = "PENDING"
-    TRAINING = "TRAINING"
-    UPLOADING_DATASET = "UPLOADING_DATASET"
-
-    def __str__(self) -> str:
-        return str(self.value)
