--- trainer/trainer/api_client/models/record_status_update_response_200_status_type_1.py
@@ -1,8 +1,0 @@
-from enum import Enum
-
-
-class RecordStatusUpdateResponse200StatusType1(str, Enum):
-    ERROR = "error"
-
-    def __str__(self) -> str:
-        return str(self.value)
